Fixed the "Prefix Information Option Processing, On-link Flag" test case.
Allow the source of the neighbour solictation to be either the
link-local or global address of the target NUT. The TAHI test
suite seems to accept link-local. RFC4861 also states that the
source address, prompting the NS, SHOULD be used as the source
of the NS. It is acceptable to use any valid address on that
interface.