Fixed the On-link Prefix does not cover TN2 test.
Modified the test so that the initial router advertisments are not sent,
as these contained source link-layer addresses, so the NUT would
never try to resolve using a neighbor solicitation. The test now
manually sends a router advertisment, which is missing link-layer
information. Updated the expected source and destination addresses of
the neighbor solicitation, so they cover the actual packets that can
legimately be receieved. Disabled neighbor discovery, so the link-layer
information is not accidentally supplied to the client too soon.